+++ src/lib/FHCollector.h
@@ -29,7 +29,7 @@ public:
   // collector functions
   void collectUString(unsigned recordId, const librevenge::RVNGString &str);
   void collectMName(unsigned recordId, const librevenge::RVNGString &name);
-  void collectPath(unsigned recordId, unsigned short graphicStyle, unsigned short layer,
+  void collectPath(unsigned recordId, unsigned graphicStyle, unsigned layer,
                    const FHPath &path, bool evenodd);
   void collectXform(unsigned recordId, double m11, double m21,
                     double m12, double m22, double m13, double m23);
